+++ lib/sms_manager/error/error.rb
@@ -1,0 +1,3 @@
+module SmsManager
+  class Error < StandardError; end
+end